After all it was useful int64.

+++ hexgrid.go
@@ -185,7 +185,7 @@ func (grid *Grid) HexCorners(hex Hex) [6]Point {
 	return corners
 }
 
-func (grid *Grid) HexNeighbors(hex Hex, layers int) []Hex {
+func (grid *Grid) HexNeighbors(hex Hex, layers int64) []Hex {
 	total := (layers + 1) * layers * 3
 	neighbors := make([]Hex, total)
 	i := 0
